requirements.txt
flask
openai
python-dotenv
requests
aiohttp
pydantic
firebolt-sdk
streamlit 